Refactor vacation segment assembler into services
