+++ config/services.yaml
@@ -365,12 +365,25 @@ services:
     MagicSunday\Memories\Clusterer\Contract\DaySummaryBuilderInterface:
         alias: MagicSunday\Memories\Clusterer\DefaultDaySummaryBuilder
 
-    MagicSunday\Memories\Clusterer\DefaultVacationSegmentAssembler:
+    MagicSunday\Memories\Clusterer\Service\TransportDayExtender: ~
+
+    MagicSunday\Memories\Clusterer\Service\RunDetector:
+        arguments:
+            $minAwayDistanceKm: 120.0
+            $minItemsPerDay: 3
+
+    MagicSunday\Memories\Clusterer\Contract\VacationRunDetectorInterface:
+        alias: MagicSunday\Memories\Clusterer\Service\RunDetector
+
+    MagicSunday\Memories\Clusterer\Service\VacationScoreCalculator:
         arguments:
             $timezone: 'Europe/Berlin'
-            $minAwayDistanceKm: 120.0
             $movementThresholdKm: 35.0
-            $minItemsPerDay: 3
+
+    MagicSunday\Memories\Clusterer\Contract\VacationScoreCalculatorInterface:
+        alias: MagicSunday\Memories\Clusterer\Service\VacationScoreCalculator
+
+    MagicSunday\Memories\Clusterer\DefaultVacationSegmentAssembler: ~
 
     MagicSunday\Memories\Clusterer\Contract\VacationSegmentAssemblerInterface:
         alias: MagicSunday\Memories\Clusterer\DefaultVacationSegmentAssembler
